Make OptionParser public + interface improvements
Expose the OptionParser interface and update the return type of Parse to
a Result type that has an Error or the OptionSet and Arguments.

Also updates TOptions to TOptionSet and TValue to TOptionValue.

In preparation for exposing a public interface for building a parser for
just the option set, the TOptions typeparams have been renamed to
TOptionSet and the OptionParser type was pulled out of the Internal
namespace.